refactor: Replace existing modals with new AddEdit components for cycles, finance admins, guards, and students; update corresponding screens for consistency
